+++ src/main/java/Product.java
@@ -1,0 +1,4 @@
+public class Product {
+    String nameProduct;
+    double priceProduct;
+}
